\name{ComICS-internal}
\alias{ComICS-internal}
\alias{.fishersMethod}
\alias{.normalizeData}
\alias{.qtlAnalysis}
\alias{.create_iQTL_association_scores}
\alias{.combine_iQTL_association_scores}
\alias{.iterative_filteration}
\title{ComICS-internal}
\description{
Internal ComICS functions
}
%\usage{}
%\arguments{}
\details{
  These are not to be called by the user.
}

%\examples{}
\keyword{internal}
