\name{commons}
\alias{immgen_dat}
\alias{DCQ_mar}
\docType{data}
\title{
 Shared Immunological datasets
}
\description{
Example datasets (Reference data and marker set):

\code{immgen_dat}: An immune cell compendium, consisting of transcriptional profiles of isolated immune cell subsets, taken from various tissues, stimulations and time points (adapted from Heng et al., 2008). 
The full immgen dataset is available for download at \url{http://dcq.tau.ac.il/} or \url{http://csgi.tau.ac.il/VoCAL/} .

\code{DCQ_mar}: Preselected group of genes that likely discriminate well between the immune-cell types given in the reference data (adapted from Altboum et al., 2014).

}
\usage{
data(commons)
}
\keyword{datasets}

