\name{dcqEx}
\alias{lung_time_series_dat}
\docType{data}
\title{
  Example datasets for runnning dcq
}
\description{
Example datasets for runnning dcq (mix data):

\code{lung_time_series_dat}: RNA-seq or microarray differential gene expression profiles of a test sample compared to a reference sample (adapted from Altboum et al., 2014).
The full dataset is available for download at \url{http://dcq.tau.ac.il/} or \url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE49934} .

uses \code{DCQ_mar} and \code{immgen_dat} from commons.RData .

}
\usage{
data(dcqEx)
}
\keyword{datasets}

