% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_img}
\alias{plot_img}
\title{Plot an image from url}
\usage{
plot_img(src)
}
\arguments{
\item{src}{path to url}
}
\value{
plots picture
}
\description{
Parses a url to be either jpeg or png and finds appropriate function to plot
}
\examples{
plot_img("https://www.r-project.org/logo/Rlogo.png")
}

