% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_jpg}
\alias{plot_jpg}
\title{Plot a jpg file from url}
\usage{
plot_jpg(src)
}
\arguments{
\item{src}{url to jpg file}
}
\value{
plotted image
}
\description{
Plot a jpg file from url
}
\examples{
plot_png("https://www.r-project.org/logo/Rlogo.png")
}

