% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_png}
\alias{plot_png}
\title{Plot a png file from url}
\usage{
plot_png(src)
}
\arguments{
\item{src}{url to png file}
}
\value{
plotted image
}
\description{
Plot a png file from url
}
\examples{
plot_png("https://www.r-project.org/logo/Rlogo.png")
}

