\name{vocalEx}
\alias{lung_dat}
\alias{gBXD}
\alias{eQTL_res}
\docType{data}
\title{
  Example datasets for runnning vocal
}
\description{
Example datasets for runnning vocal (Expression data, genotype data and eQTL results data): 

\code{lung_dat}: RNA-seq or microarray gene-expression profiles of a given complex tissue across a population of genetically distinct (genotyped) individuals (adapted from E-MTAB-848).

\code{gBXD}: Genotyping of the different individuals under study (adapted from GeneNetworks).

\code{eQTL_res}: eQTL analysis results of the different genes in the expression data (specifically the genes that appear in the marker set(s) selected).

uses \code{DCQ_mar} and \code{immgen_dat} from commons.RData .

}
\usage{
data(vocalEx)
}
\keyword{datasets}