% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xkcd.R
\name{xkcd}
\alias{xkcd}
\title{Show me an XKCD comic}
\usage{
xkcd(num = NULL, print_alt = TRUE)
}
\arguments{
\item{num}{optional; which XKCD number do you want to see? Default is random}

\item{print_alt}{should we print the alt text?}
}
\value{
plots a comic
}
\description{
Plots to
}
\examples{
xkcd()
xkcd(123)
}

